+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/combobox/ComboBoxBuilder.java
@@ -52,6 +52,14 @@ public interface ComboBoxBuilder<T, C extends JComboBox<T>, B extends ComboBoxBu
 	B completionMode(Completion.Mode completionMode);
 
 	/**
+	 * Specifies whether to normalize strings during auto completion
+	 * @param normalize true if strings should be normalized during autocomplete
+	 * @return this builder instance
+	 * @see #completionMode(Completion.Mode)
+	 */
+	B normalize(Completion.Normalize normalize);
+
+	/**
 	 * @param renderer the renderer for the combo box
 	 * @return this builder instance
 	 * @see JComboBox#setRenderer(ListCellRenderer)